Fix tests in Travis CI
It seems the preferred method for installing gems in Travis CI is now
using bundler. To fix this a Gemfile has been added that uses the
gemspec file to install the necessary gems. This causes problems since
Hoe doesn't use a gemspec file and just builds a package. Add
hoe-gemspec to generate a gemspec file for the Gemfile to use.